Calculate off-diagonal elements correctly
Previously, the admittance between buses was calculated to include the shunt admittance of the line. This was incorrect, since the shunt admittance is between a bus and ground. Now the admittance between buses includes only the inverse of the series impedance of the line. The resulting Ybus matrix matches what PowerWorld calculates.